+++ adapters/InMobi/InMobiAdapter/GADMAdapterInMobi.h
@@ -15,7 +15,6 @@
 
 @interface GADMAdapterInMobi
     : NSObject <GADMAdNetworkAdapter, IMBannerDelegate, IMInterstitialDelegate, IMNativeDelegate>
-+ (BOOL)isAppInitialised;
 @property(nonatomic, retain) IMBanner *adView;
 @property(nonatomic, retain) IMInterstitial *interstitial;
 @property(nonatomic, retain) IMNative *native;